Adding support to builds for mac os x on the installation scripts and general support for profile updating, venue registration and more

+++ include/backend/models/Venue.hpp
@@ -15,8 +15,8 @@ namespace indiepub
     public:
         Venue() = default;
         Venue(const std::string &venue_id, const std::string &owner_id, const std::string &name,
-              const std::string &location, int capacity, std::time_t created_at);
-        
+              const std::string &location, long capacity, std::time_t created_at);
+
         static const std::string COLUMN_FAMILY;
         static const std::string IDX_VENUES_NAME;
         static const std::string IDX_OWNERS_ID;
@@ -26,7 +26,7 @@ namespace indiepub
         std::string owner_id() const;
         std::string name() const;
         std::string location() const;
-        int capacity() const;
+        long capacity() const;
         std::time_t created_at() const;
 
         // JSON serialization
@@ -41,7 +41,7 @@ namespace indiepub
         std::string owner_id_; // UUID (links to User)
         std::string name_;
         std::string location_;
-        int capacity_;
+        long capacity_;
         std::time_t created_at_;
     };
 }
